server/requirements.txt
flask
flask-restful
requests
flask-script
flask-cors
flask-sqlalchemy
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
celery[redis]
celery
bcrypt
redis
Flask-JWT-Extended

gunicorn
pytest